DVX-6600 Move common functions to its own file
And also move its tests.
