+++ pkg/dnsbl/common.go
@@ -1,0 +1,21 @@
+package dnsbl
+
+import (
+	"strings"
+)
+
+// Reverse reverses slice of string elements.
+func reverse(original []string) {
+	for i := len(original)/2 - 1; i >= 0; i-- {
+		opp := len(original) - 1 - i
+		original[i], original[opp] = original[opp], original[i]
+	}
+}
+
+// ReverseAddress converts IP address in string to reversed address for query.
+func reverseAddress(ipAddress string) (reversedIPAddress string) {
+	ipAddressValues := strings.Split(ipAddress, ".")
+	reverse(ipAddressValues)
+	reversedIPAddress = strings.Join(ipAddressValues, ".")
+	return
+}
