+++ pkg/dnsbl/dnsbl.go
@@ -6,7 +6,6 @@ import (
 	"io"
 	"log"
 	"net"
-	"strings"
 	"sync"
 )
 
@@ -78,20 +77,4 @@ func (c *Checker) query(address string) int {
 		log.Printf("%v returned %v\n", address, result)
 	}
 	return len(result)
-}
-
-// Reverse reverses slice of string elements.
-func reverse(original []string) {
-	for i := len(original)/2 - 1; i >= 0; i-- {
-		opp := len(original) - 1 - i
-		original[i], original[opp] = original[opp], original[i]
-	}
-}
-
-// ReverseAddress converts IP address in string to reversed address for query.
-func reverseAddress(ipAddress string) (reversedIPAddress string) {
-	ipAddressValues := strings.Split(ipAddress, ".")
-	reverse(ipAddressValues)
-	reversedIPAddress = strings.Join(ipAddressValues, ".")
-	return
 }
